Esquema,datos sql cambiados y entidad studio modificado, revisar por si acaso

+++ src/main/java/org/iesalixar/daw2/Alejandroangulomendez/game_zone/repositories/StudioRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
  * Repositorio para la entidad Studio que extiende JpaRepository.
  * Proporciona operaciones CRUD y consultas personalizadas para la entidad Studio.
  */
-public interface StudioRepository extends JpaRepository<Studio, Long> {
+public interface StudioRepository extends JpaRepository<Studio, Integer> {
 
     /**
      * Obtiene todos los estudios.
@@ -34,7 +34,7 @@ public interface StudioRepository extends JpaRepository<Studio, Long> {
      *
      * @param id el ID del estudio a eliminar.
      */
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     /**
      * Obtiene un estudio por su ID.
@@ -42,7 +42,7 @@ public interface StudioRepository extends JpaRepository<Studio, Long> {
      * @param id el ID del estudio.
      * @return un Optional que contiene el estudio si se encuentra, o vacío si no se encuentra.
      */
-    Optional<Studio> findById(Long id);
+    Optional<Studio> findById(Integer id);
 
     /**
      * Comprueba si existe un estudio con un nombre específico.
@@ -60,5 +60,5 @@ public interface StudioRepository extends JpaRepository<Studio, Long> {
      * @return true si existe un estudio con el nombre especificado (excluyendo el estudio con el ID dado), false en caso contrario.
      */
     @Query("SELECT COUNT(s) > 0 FROM Studio s WHERE s.name = :name AND s.id != :id")
-    boolean existsStudioByNameAndNotId(@Param("name") String name, @Param("id") Long id);
+    boolean existsStudioByNameAndNotId(@Param("name") String name, @Param("id") Integer id);
 }
